move DefaultThumbnail to separate file and BookDetails to separate directory
